+++ ios/CleverTapReact/CleverTapReact.mm
@@ -704,6 +704,13 @@ RCT_EXPORT_METHOD(dismissInbox) {
     [[self cleverTapInstance] dismissAppInbox];
 }
 
+RCT_EXPORT_METHOD(fetchInbox:(RCTResponseSenderBlock)callback) {
+    RCTLogInfo(@"[CleverTap fetchInbox]");
+    [[self cleverTapInstance] fetchInboxWithCallback:^(BOOL success) {
+        [self returnResult:@(success) withCallback:callback andError:nil];
+    }];
+}
+
 RCT_EXPORT_METHOD(initializeInbox) {
     RCTLogInfo(@"[CleverTap Inbox Initialize]");
     [[self cleverTapInstance] initializeInboxWithCallback:^(BOOL success) {
@@ -854,6 +861,11 @@ RCT_EXPORT_METHOD(pushDisplayUnitViewedEventForID:(NSString*)unitId) {
 RCT_EXPORT_METHOD(pushDisplayUnitClickedEventForID:(NSString*)unitId) {
     RCTLogInfo(@"[CleverTap pushDisplayUnitClickedEventForID]");
     [[self cleverTapInstance] recordDisplayUnitClickedEventForID:unitId];
+}
+
+RCT_EXPORT_METHOD(pushDisplayUnitElementClickedEventForID:(NSString*)unitId elementID:(NSString*)elementId additionalProperties:(NSDictionary*)additionalProperties) {
+    RCTLogInfo(@"[CleverTap pushDisplayUnitElementClickedEventForID]");
+    [[self cleverTapInstance] recordDisplayUnitElementClickedEventForID:unitId elementID:elementId additionalProperties:additionalProperties];
 }